[Refactor]#107-VM에 스캔 관련 함수 이동
